Introduce retworkx-core crate (#459)
* Introduce retworkx-lib crate

This commit adds a new workspace crate to retworkx, retworkx-lib, which
contains a pure rust library built on top of petgraph for retworkx
algorithms and additional non-python specific data structures. Prior to
this we didn't really have a public rust API for retworkx as the main
output from the retworkx crate was a cdynlib for use with python. By
creating a new crate this gives us a place to enable rust reuse of the
library for functionality that is not specific to python. Any potential
functionality in retworkx-lib will need to be solely reusable in rust
without pyo3 or python.

The primary retworkx crate will still be primarily for the python
library. But for places where it makes sense to expose a rust api we can
use the retworkx-lib crate to export those for other rust usage.

Closes #108

* Fix setuptools-rust builds

The MANIFEST.in was missing the retworkx-lib crate directory which meant
when setuptools built the sdist it wasn't including the source for the
crate. So when setuptools-rust was called on building the sdist it
errors because it can't find the Cargo.toml for retworkx-lib or any of
the source.

* Migrate max weight matching to retworkx-lib

* Add k_shortest_path module to retworkx-lib

* Add README

* Start adding api documentation

* Add rust tests for retworkx-lib to CI

* Add retworkx-lib to the contributing guide

* Migrate dfs_edges() to retworkx-lib

* Remove unused badges section

* Improve documentation and rework lib structure slightly

* Add retworkx-lib doc builds to ci

* Fix lint

* minor doc changes

* Tighten capacity for dfs_edges allocation

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Make min_scored private

* Run rustfmt

* Make k_shortest_path score value generic

* Mention cargo doc --open in CONTRIBUTING.md

* Tweak overview wording

* Apply suggestions from code review

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Bump ahash version

* Add result infallible alias type for doc tests

* Make weight_callable() generic for return type

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Rename retworkx-lib to retworkx-core

* Update retworkx-core/Cargo.toml

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>

Co-authored-by: georgios-ts <ts.giorgos@yahoo.gr>
Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>