Restore subTest usage in max_weight_matching tests
This commit restores the use of unittest subTests for each graph in the
random graph max_weight_matching() tests. The max_weight_matching()
tests uses testtools (a unittest extension library) instead of vanilla
unittest to enable stdout capture in the those tests (as some tests
emit stdout to inform us of alternative valid results that don't match
networkx's output) which unittest doesn't offer. However the previous
releases of testtools (< 2.5.0) had an incompatibility with stdlib
unittest's subTest functionality that caused errors. At the time the
subTest usage was removed to enable the stdout capture. Now that
testtools 2.5.0 has been out for several months we can restore the
subTest usage so that error messages say which input we ran with.
This commit makes that change and also sets the minimum testtools
version to the latest release 2.5.0 in the tox.ini to ensure we are
using it.