Remove travis configuration file
This commit removes the travis configuration file from the repo. For
roughly 6 months now we haven't been using travis for CI on retworkx
mostly because of billing changes and tight quotas for open source
projects means it is not a realistic option for free use by open
source projects anymore. After the recent security incidents on
travis is no longer enabled on the repository and while it was the only
CI service offering a native non-x86 environment it is unfortunately not
a viable solution anymore. However, the configuration file remained in
the repository even though we're no longer using it. This commit just
cleans up the stray file as it's no longer needed.