Add online topological sorter.
This PR implements a new class `TopologicalSorter` to topologically sort a directed graph.
It gives the ability to process nodes while they are becoming ready (i.e they have no unprocessed
predecessors) and then mark them as done, so more nodes can be freed.

`TopologicalSorter` wraps the underlying graph inside a `Py<..>` with GIL-independent lifetime
in order to avoid cloning the graph. The downside is that users may modify the graph while using
a `TopologicalSorter`, we do not guard against it and this might result in wrong outputs.