Try enabling isolated_build option
CI is failing without the pip install fallback because tox is calling
`python setup.py sdist` to build the sdist prior to installing. But this
isn't processing the pyproject.toml because nothing is checking for it.
In an attempt to make tox try and respect the pyproject.toml this turns
on the isolated_build option which creates a venv for building which
hopefully will mean tox looks at the pyproject.toml prior to running
`python setup.py sdist`.