Explicitly list setuptools as the build system in pyproject.toml
Using build isolation tox is failing because we don't explicitly list
setuptools as the build system backend in the pyproject.toml. This isn't
an issue with pip because it implicitly uses setuptools as the backend
(although via a legacy interface) if the build-backend isn't specified.
However, tox doesn't have this default and requires the build system
backend be specified. This commit corrects this oversight and explicitly
lists setuptools as the build-backend field in the pyproject.toml.