Remove optional dependency on pydot (#534)
* Remove optional dependency on pydot

Previously we were relying on the pydot library to call graphviz to
generate visualizations. This was done for simplicity since pydot is all
about providing a python API wrapper around graphviz. However, it does
this by parsing the dot files into it's own internal representation of
the graph description to provide a python api for interacting with it
prior to calling out to graphviz. For retworkx this is unnecessary as we
have a native function for generating dot files and we just need to pass
that directly to graphviz. Additionally, this parsing adds a place for
bugs to be introduced as was seen in #526 when a new version of
pyparsing (the dependency pydot uses to parse the dot files) started
causing invalid output to be generated. This commit updates the graphviz
drawer to stop using pydot and instead manually calling graphviz
directly with the retworkx generated dot files. The setup.py's optional
extras are then updated to no longer list pydot as required for the
graphviz drawer. This should hopefully reduce the complexity of the code
path and avoid potential future issues.

Fixes #526

* Remove missed references to pydot

This commit removes missed references to pydot in the README and the
unittests.