Add edge utility methods to PyGraph and PyDiGraph (#538)
* Add edge utility methods to PyGraph and PyDiGraph

There were some gaps in working with edges in the graph class API.
Mainly when working with indices there was no method to acquire the
edge's data payload or endpoints easily. Additionally for the PyGraph
class didn't have a simple method to find incident edges directly
instead a user had to rely on neighbors() and get the edges from that
directly. This commit adds methods to fill these API gaps to make
working with edges a bit easier.

* Add more methods

* Fix docstring for new methods

* Apply suggestions from code review

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Add all_edges kwarg to to PyDigraph.incident_edges()

This commit adds a flag to the directed graph version of
incident_edges() to enable users to get both the incoming and outgoing
edges if they so desire.

* Apply suggestions from code review

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Add incident_edge_index_map method

* Assert endpoint order for PyGraph.incident_edge_index_map()

Previously, the endpoint order for PyGraph.incident_edge_index_map()
wasn't guaranteed. In practice it appears that the source node was
always the user provided node, but that wasn't actually defined
behavior. This commit updates the internal petgraph calls to ensure we
always return the endpoints in a fixed order and adds that to the
documented behavior of the method.

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>