Documentation fixes and improvements
This commit goes through the documentation for retworkx and fixes
various typos and mistakes. It also updates the generated api docs to be
put in an 'apiref' path instead of 'stubs' to clean up the urls a bit.
To maintain compatibility with existing links though a redirect is added
to redirect from stubs/* to apiref/*.

Fixes #507