Add dfs-search (#456)
* Add dfs-search

Widely used graph libraries like Boost Graph Library and graph-tool
provide the ability to insert callback functions at specified event
points for many common graph search algorithms. petgraph has a similar
concept in `petgraph::visit::depth_first_search` function. This commit
implements an iterative version of `depth_first_search`, that will be
used in a follow-up in the pending PRs #444, #445. At the same time it
exposes this new functionality in Python by letting users subclassing
`retworkx.visit.DFSVisitor` and provide their own implementation for
the appropriate callback functions.

The benefit is less memory consumption since we avoid storing the results
but rather let the user take the desired action at specified points. For
example, if a user wants to process the nodes in dfs-order, we don't need
to create a new list with all the graph nodes in dfs-order but rather the user
can process a node on the fly. We can (probably) leverage this approach
in other algorithms as an alternative for our inability to provide "real" python
iterators.

* break or prune the search tree with custom exceptions

* tests

* move `depth_first_search` to retworkx-core + create a new traversal module together with `dfs_edges` function

* add option to pass a vector of starting nodes + in dfs events report the weight of an edge

* lint

* fix docs

* this is the last fix

* define custom exceptions `StopSearch`, `PruneSearch` in python and import them in rust so we can simplify a bit the namespaces

* minor doc fixes

* ignore flake warning

* run black

* deduplicate macro and mention petgraph DfsEvent struct