DNM: Test wheel jobs after changes
This commit should not be merged without being reverted. This commit
changes the ci configuration to run on PRs only and will not ublock the
built wheels to pypi. This is being used to validate the changes made in
earlier commits and will be used to iterate on any fixes necessary until
the new configuration works.