Skip musllinux on slow cross build jobs
The slow cross build jobs have custom skips set to split the builds
across multiple jobs. However those skip lists were not updated to
exclude musl libc linux builds like the general config. This was causing
a failure because we don't support musl libc currently and the build
scripts don't work in that build environment. This commit adds the
missing skip to the job configurations where it was missing to fix this
issue.