Set max line length to 100 characters
Previously we had been setting rustfmt and black to use an 80 character
line length. This was primarily because my personal dev workflow is
setup around 80 character-wide windows. However, as the project has a
growing number of contributors my personal archaic taste in dev workflow
should not be subjected upon everyone. This commit bumps the max line
length in rustfmt and black up to 100 which is a more common default
that will work better in most editors and not needlessly wrap lines
everywhere. Most of this diff is just the auto-formatting adjustments
now that there are an additional 20 characters available for them to
use.

Additionally, with the recent release of black 22.0 the project has
moved out of beta and adopted a stability policy. [1] Moving forward for
a given major version there will be no formatting changes in black, so
this commit bumps the black version and loosens the constraint to use
any of the 22.x.y which means we'll get the latest bug fixes for running
in CI but have a stable set of formatting rules applied.

[1] https://black.readthedocs.io/en/latest/the_black_code_style/index.html#stability-policy