Upgrade PyO3 and numpy to 0.16.2 (#569)
* Upgrade PyO3 and numpy to 0.16.0

This commit updates pyo3 and numpy to the latest release 0.16.0. The
details on these releases can be found here:

https://pyo3.rs/v0.16.0/changelog.html

and

https://github.com/PyO3/rust-numpy/blob/main/CHANGELOG.md

Unfortunately there are a number of breaking changes and deprecations we
will have to update our usage to account for. This commit starts the
process by fixing the compiler errors due to changes in behavior a
subsequent commit will address the deprecation warnings.

* fix pyproto deprecation warnings

* Fix error making coverage fail

* Bump PyO3 to 0.16.1

* Bump to latest pyo3 and numpy release

* Update additional dependencies in Cargo.lock

Co-authored-by: georgios-ts <ts.giorgos@yahoo.gr>
Co-authored-by: Ivan Carvalho <ivancarv@student.ubc.ca>