Add return all simple paths between all pairs of nodes
This commit adds a new function all_pairs_all_simple_paths
which is used to return all simple paths between all nodes of a
particular length. It basically is a multithreaded loop around the
existing all_simple_paths function that runs the function between all
pairs of nodes in the graph.