Update workspace configuration to be tied to specific versions (#583)
This commit updates the retworkx-core crate dependency to have a fixed
version which matches the retworkx crates version. We're not going to be
in a situation where we we release retworkx-core separately from the
main retworkx crate we only want to support the main crate with the core
crate at the same version. This commit makes this relationship
explicitly. At the same time it fixes the version string in the
retworkx-core crate as it wasn't updated in #560 like it should have
been.

(cherry picked from commit a0aaa1f93ef21fe77dccee63cd37fdb8f43c7cee)