Add section to intro tutorial on graph element modification (#596)
This commit adds a section to the introduction tutorial on graph element
modification. It's a basic feature of interacting with a graph to show
how you can update a node or edge in place on an existing graph object
however this was previously not mentioned in the introduction tutorial.
This commit corrects this oversight by explaining how this can be done.

(cherry picked from commit d02b5446d91cb38a117d18482f97f2b27d9e08e3)