Add support for slice inputs to custom sequence return types (#591)
* Add support for slice inputs to custom sequence return types

This commit adds support for dealing with slice inputs to custom
sequence return types such as NodeIndices. Previously, if a slice
was requested from a custom return type it would have raised a
TypeError. With this commit now it will return a new container object
with a copy of the data from the requested slice.

Fixes #590

* Fix windows build

* Fix handling of negative length int index

Previously if an integer index was passed to `__getitem__`
with a value equal to the negative length of the sequence
we would incorrectly raise an except when instead the first
element should have been returned. This commit fixes the
handling of that edge case.

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Fix handling of negative steps in slices

This commit fixes the handling of negative slices in sequence custom
return types. Previouly using a negative step would cause a panic
instead of correctly returning a reversed slice from the container. This
commit corrects this so we can deal with negative slice steps too.

* Apply suggestions from code review

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>