Fix generic usage for some retworkx-core functions
This commit fixes several bugs with the use of generics bugs in
retworkx-core were fixed for max weight matching and betweenness
centrality as they both were only implementable previously with graphs
that used the default Graph/StableGraph u32 NodeIndex type. This was an
artificial bound on the input type as the functions should work with any
valid index type. This commit removes that bound so that functions work
with any indexable graph.