Optimize performance of some generator functions (#625)
* Optimize performance of some generator functions

This commit makes some improvements to the internals cycle graph, path
graph, star graph, and  mesh graph generator functions to optimize the
performance and reduce the memory overhead when creating the graphs. The
improvements fall into a couple categories, first the graph memory
allocation is only done once now as opposed to on demand as we create
the graph, and second intermediate index lists are no longer used in
places they weren't required to avoid the overhead of having to maintain
dual Vecs.

Related to #623

* Further simplify generators

This commit builds off the first one and removes more intermediate Vecs
for storing node indices and also reduces the number of allocations on
more functions.

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Fix zero length handling in modified generators

* Fix potential overflow in 0 length graphs

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>