Rename retworkx to rustworkx
This commit renames the retworkx project to rustworkx. The maintainers of
NetworkX requested that we change the name of the library because they
viewed networkx as being too similar to retworkx. Out of respect for
this request and the NetworkX library which inspired retworkx we are
renaming the library to rustworkx. Rustworkx is a similar name to
retworkx and enables us to keep using the rx abbreviation for imports
which is commonly used.

This commit, migrates all the internal usage and names from retworkx to
rustworkx. This includes the retworkx-core crate which has similarly been
renamed to rustworkx-core. However, for backwards compatibility and to
ease the migration of our existing user base the retworkx package is
still included in rustworkx. An import redirect is added so that
import retworkx.foo will return return rustworkx.foo. Additionally, the
README and documentation are updated to mention the rename.

The migration plan this commit takes is that for the pending 0.12.0
release we'll migrate to rustworkx for everything but keep the backwards
compatibility shim in place with no warning. In the 0.13.0 release we'll
emit a DeprecationWarning (or a FutureWarning) that the name is
deprecated (this waits a release to ensure there is upgrade path where
> 1 version works without raising a warning to ease migration). We can
then remove the deprecated retworkx name as part of the 1.0.0 release.
During the transition period we'll continue to publish releases to both
the rustworkx and retworkx names through at least 0.14.0 (but with the option
to keep publishing both until 1.0.0) so that users will see the deprecation
warnings when importing retworkx.

Closes #641