Revert "Make rustworkx and retworkx independent dependent packages"
It turns out that adding a dependency is not going to work prior to the
rustworkx package being released because pip won't be able to resolve
the rustworkx dependency. This commit reverts this structure for the
time being. Closer to the 0.12.0 release we can revisit this and plan to
do this structure (maybe by releasing rustworkx first, making this
change and then releasing retworkx with the dependency).

This reverts commit 2ae3ffd63a2ceecf2d5aedde62952d1014719ed9.