Rename retworkx to rustworkx (#644)
* Rename retworkx to rustworkx

This commit renames the retworkx project to rustworkx. The maintainers of
NetworkX requested that we change the name of the library because they
viewed networkx as being too similar to retworkx. Out of respect for
this request and the NetworkX library which inspired retworkx we are
renaming the library to rustworkx. Rustworkx is a similar name to
retworkx and enables us to keep using the rx abbreviation for imports
which is commonly used.

This commit, migrates all the internal usage and names from retworkx to
rustworkx. This includes the retworkx-core crate which has similarly been
renamed to rustworkx-core. However, for backwards compatibility and to
ease the migration of our existing user base the retworkx package is
still included in rustworkx. An import redirect is added so that
import retworkx.foo will return return rustworkx.foo. Additionally, the
README and documentation are updated to mention the rename.

The migration plan this commit takes is that for the pending 0.12.0
release we'll migrate to rustworkx for everything but keep the backwards
compatibility shim in place with no warning. In the 0.13.0 release we'll
emit a DeprecationWarning (or a FutureWarning) that the name is
deprecated (this waits a release to ensure there is upgrade path where
> 1 version works without raising a warning to ease migration). We can
then remove the deprecated retworkx name as part of the 1.0.0 release.
During the transition period we'll continue to publish releases to both
the rustworkx and retworkx names through at least 0.14.0 (but with the option
to keep publishing both until 1.0.0) so that users will see the deprecation
warnings when importing retworkx.

Closes #641

* Make rustworkx and retworkx independent dependent packages

This commit splits out the retworkx component from the rustworkx
package. This makes the new rustworkx package solely contain the
rustworkx piece and the retworkx piece becomes an independent package
that depends on rustworkx (and will internally redirect everything to
rustworkx).

Co-authored-by: Ivan Carvalho <ivancarv@student.ubc.ca>

* Update lint doc job definition for new package name

* Revert "Make rustworkx and retworkx independent dependent packages"

It turns out that adding a dependency is not going to work prior to the
rustworkx package being released because pip won't be able to resolve
the rustworkx dependency. This commit reverts this structure for the
time being. Closer to the 0.12.0 release we can revisit this and plan to
do this structure (maybe by releasing rustworkx first, making this
change and then releasing retworkx with the dependency).

This reverts commit 2ae3ffd63a2ceecf2d5aedde62952d1014719ed9.

* Revert "Revert "Make rustworkx and retworkx independent dependent packages""

This reverts commit 1d04a30bda7308b23171368af9a8a0e2478794fa.

* Remove unsupported set command

* Use coreutils to set environment variable on Windows

* Use coreutils command in .cargo/bin

* Fix Sphinx warnings

* Update tox.ini

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

Co-authored-by: Ivan Carvalho <ivancarv@student.ubc.ca>
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>