Improve introductory documentation materials (#657)
* Improve introductory documentation materials

This commit improves some of the introductory material for the rustworkx
documentation. The introductory guides for rustworkx were a bit lacking,
partially because it inherited the basic install guide from the readme
and also because the introduction tutorial didn't cover the very
beginning of what rustworkx is or how to use it. This commit addresses
this by adding a dedicated install/getting started page which includes
a platform support table and adding a couple sections to the
introduction tutorial to cover basic installation and importing.
Additionally, the imports in the documentation are standardized on
aliasing rustworkx as rx.

This commit superscedes #559 which added the platform support table to
the readme.

* Fix sphinx labels

* Fix typo

* Update docs/source/install.rst

* Apply suggestions from code review

* Update docs/source/index.rst

* Remove stale README.rst

* Add missing label