Add support for Python 3.11
For the upcoming 0.12.0 release it would be good to add support for
python 3.11. While this hasn't been released yet it's currently a
release candidate with a frozen C API. We're likely to release rustworkx
0.12.0 prior to the Python 3.11 release. In order to avoid having to
have rustworkx be a blocker for any downstream packages looking to adopt
3.11 this commit proactively enables support for testing and building
with Python 3.11 off the latest rc in order to ensure we're building
binaries compatible with 3.11 for the 0.12.0 release.