Pin Rust to 1.63 on linux wheel jobs
In the upcoming Rust 1.64.0 release the minimum glibc version is being
raised to 2.17 which is higher than the minimum supported version in
manylinux 2010 which is our current Python wheel target environment. To
continue supporting manylinux 2010 for Rustworkx 0.12.0 we need to pin
the rust version to avoid implicitly increasing the minimum after rust
1.64 is released. This also adds a release note documenting that this is
the last release that supports manylinux 2010.